refactor: break runtime import cycles via leaf-module imports
