+++ src/synthorg/communication/meeting/config.py
@@ -1,12 +1,15 @@
 """Meeting protocol configuration models (see Communication design page)."""
 
-from pydantic import BaseModel, ConfigDict, Field
+from typing import Self
+
+from pydantic import BaseModel, ConfigDict, Field, model_validator
 
 from synthorg.communication.meeting.enums import (
     ConflictDetectorType,
     MeetingProtocolType,
 )
-from synthorg.core.types import NotBlankStr
+from synthorg.communication.meeting.frequency import MeetingFrequency
+from synthorg.core.types import NotBlankStr, validate_unique_strings
 
 
 class RoundRobinConfig(BaseModel):
@@ -168,3 +171,81 @@ class MeetingProtocolConfig(BaseModel):
         default_factory=StructuredPhasesConfig,
         description="Structured-phases protocol settings",
     )
+
+
+class MeetingTypeConfig(BaseModel):
+    """Configuration for a single meeting type.
+
+    Maps to the Communication design page ``meetings.types[]``.  Exactly one of
+    ``frequency`` or ``trigger`` must be set.
+
+    Attributes:
+        name: Meeting type name (e.g. ``"daily_standup"``).
+        frequency: Recurrence schedule (mutually exclusive with trigger).
+        trigger: Event trigger (mutually exclusive with frequency).
+        participants: Participant role or agent identifiers.
+        duration_tokens: Token budget for the meeting.
+    """
+
+    model_config = ConfigDict(frozen=True, allow_inf_nan=False, extra="forbid")
+
+    name: NotBlankStr = Field(description="Meeting type name")
+    frequency: MeetingFrequency | None = Field(
+        default=None,
+        description="Recurrence schedule",
+    )
+    trigger: NotBlankStr | None = Field(
+        default=None,
+        description="Event trigger",
+    )
+    participants: tuple[NotBlankStr, ...] = Field(
+        default=(),
+        description="Participant role or agent identifiers",
+    )
+    duration_tokens: int = Field(
+        default=2000,
+        gt=0,
+        description="Token budget for the meeting",
+    )
+    protocol_config: MeetingProtocolConfig = Field(
+        default_factory=MeetingProtocolConfig,
+        description="Meeting protocol configuration",
+    )
+    min_interval_seconds: int | None = Field(
+        default=None,
+        ge=1,
+        description="Minimum seconds between event-triggered meetings of this type",
+    )
+
+    @model_validator(mode="after")
+    def _validate_frequency_or_trigger(self) -> Self:
+        """Exactly one of frequency or trigger must be set.
+
+        Returns:
+            The validated meeting-type config.
+
+        Raises:
+            ValueError: If both or neither of ``frequency`` / ``trigger``
+                are set, or ``min_interval_seconds`` is set without a
+                trigger.
+        """
+        if self.frequency is not None and self.trigger is not None:
+            msg = "Only one of frequency or trigger may be set, not both"
+            raise ValueError(msg)
+        if self.frequency is None and self.trigger is None:
+            msg = "Exactly one of frequency or trigger must be set"
+            raise ValueError(msg)
+        if self.min_interval_seconds is not None and self.trigger is None:
+            msg = "min_interval_seconds requires trigger-based meetings"
+            raise ValueError(msg)
+        return self
+
+    @model_validator(mode="after")
+    def _validate_participants(self) -> Self:
+        """Ensure participant entries are unique.
+
+        Returns:
+            The validated meeting-type config.
+        """
+        validate_unique_strings(self.participants, "participants")
+        return self
